Moved modal utilities file into discussion modal utilities since they really only work for discussion related things anyway
