+++ src/pieces/discussion/menu/staff/managePostScoring/PostScoringModals.ts
@@ -1,12 +1,11 @@
-import { ActionRowBuilder, ButtonInteraction, CacheType, Client, ModalSubmitInteraction, TextInputBuilder, TextInputStyle } from "discord.js";
+import { ActionRowBuilder, ButtonInteraction, Client, ModalSubmitInteraction, TextInputBuilder, TextInputStyle } from "discord.js";
 import { AwardSpecs, PostSpecs } from "../../../../../generalModels/DiscussionScoring";
 import { DATABASE_ERROR_MESSAGE, getCourseByName, overwriteCourseDiscussionSpecs } from "../../../../../generalUtilities/CourseUtilities";
 import { DEFAULT_POST_SPECS } from "../../../../../pieces/courseManagement/DiscussionRulesDefaults";
-import { DiscussionModalData, DiscussionModalHandler, createDiscussionModal } from "../../../../../pieces/menu/ModalUtilities";
-import { recollectManagePostSpecsInput, refreshManagePostSpecsMenu } from "./ManagePostScoringMenu";
-import { INPUT_ERROR_PREFIX, INVALID_INPUT_PREFIX, SCORING_ERROR_MESSAGE } from "../../DiscussionModalUtilities";
-import { rescoreDiscussion } from "../../../../../pieces/discussion/scoring/rescorePeriods";
 import { CourseQueue } from "../../../../../pieces/discussion/scoring/courseQueue";
+import { rescoreDiscussion } from "../../../../../pieces/discussion/scoring/rescorePeriods";
+import { DiscussionModalData, DiscussionModalHandler, INPUT_ERROR_PREFIX, INVALID_INPUT_PREFIX, SCORING_ERROR_MESSAGE, createDiscussionModal } from "../../DiscussionModalUtilities";
+import { recollectManagePostSpecsInput, refreshManagePostSpecsMenu } from "./ManagePostScoringMenu";
 
 // POST SCORE INPUT COMPONENT
 const SCORE_INPUT_ID = "discussion_score_input";
